mysql to postgresql

+++ memberserver/src/main/scala/com/jxjxgo/membercenter/domain/Member.scala
@@ -5,4 +5,4 @@ import java.sql.Timestamp
 /**
   * Created by fangzhongwei on 2017/1/4.
   */
-case class Member(memberId: Long, mobile: String, mobileTicket: String, status: Byte, nickName: String, password: String, gmtCreate: Timestamp, gmtUpdate: Timestamp)
+case class Member(memberId: Long, mobile: String, mobileTicket: String, status: Short, nickName: String, password: String, gmtCreate: Timestamp, gmtUpdate: Timestamp)
